make lkl config json embedded at build time
In lkl, network configs are specified by json file at runtime. rexec is
responsible to open and create fd for the config. However, solo5 does
not allow to supply such external files. To overcome this issue, let
platform specific code embed the config using gen-config.sh. This
solution is not so flexible, it will be changed in future commmits to
load at runtime.

Signed-off-by: Akira Moroo <retrage01@gmail.com>